Added celsius and fahrenheit links
changed the labeling of the "celsius" and "fahrenheit" links in html from <span id=""> to <a href="#" id:""> because they're technically links even though the page doesnt change. Added windspeed and humidity. Make sure EVERYTHING IS SPELLED CORRECTLY. to make sure that the information from js shows up on the webpage