Added js formula for forecast
added the basic info needed for daily forecast aka temp and day(but the days are currently "undefined") for 5 days but all the info is just the same info copied and pasted. Need help defining each days temp and accurate date. Also. icon using icons that define the weather of that day accurately.