Update index.js
havent done it yet but i may have to get rid of formula and do long hand to show the temperature for each day bc the temps i want aren't one after the other. The api shows every 3 hours, i think i want to show the temp. everyday at noon or 3pm.
Changed the icon url into a variable so that the icon pics matched the information aka weather for the forecast.