current weather icon is functional
rearranged and resized many things in css. added bootstrap so columns are active now. want the app to be rectangular, with the horizontal side the longer(spread across instead of scrolling up and down. So making everything smaller. All the icons are accurate but it's in 3hr rotation rather than daily(next step). The time variable is not active yet(next step). Then i will focus on visuals, changing the icons to another style and introducing color and fonts.